build: Add MPIF08 configure check for Intel MPI
Adds an additional configure check to check if Intel MPI
fails to build a simple MPIF08 test case. This is due to a
bug in Intel MPI which has to be fixed after installation of Intel
MPI, requiring to rebuild the MPIF08 bindings.

If not checked, builds with Intel MPI will fail due to the bug
described in:

  https://community.intel.com/t5/Intel-MPI-Library/MPI-f08-with-polymorphic-argument-CLASS/m-p/1590421

When this bug is triggered, disable the entire MPIF08 support.
Steps on how to solve this are already described in OPEN_ISSUES.

Signed-off-by: Jan André Reuter <j.reuter@fz-juelich.de>